requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
chess==1.10.0
click==8.1.8
colorama==0.4.6
dpkt==1.9.8
Flask==1.1.2
idna==3.10
importlib-metadata==6.7.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.1.5
Pillow==9.5.0
PyCap==1.1.3
pynput==1.8.1
python-chess==1.999
requests==2.31.0
scapy==2.6.1
semantic-version==2.10.0
six==1.17.0
typing-extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.0.3
winpcapy==1.0.3
zipp==3.15.0
flask
gunicorn
python-dotenv

